+++ info_projekt/lib/services/firestore_service.dart
@@ -7,9 +7,7 @@ class FirestoreService {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
 
   //DateTime registrationDate = DateTime.now();
-  num balance = 0;
   String iban = 'null';
-  num tax_pot = 0;
 
   Future<void> saveUserDataToFirestore(
       String userName, String registrationDate) async {
@@ -19,31 +17,13 @@ class FirestoreService {
     Map<String, dynamic> datatoSave = {
       'email': userName,
       'registrationDate': DateTime.now(),
-      'iban': iban,
-      'balance': balance,
-      'taxpot': tax_pot,
       'UID': UID
     };
-
-    Map<String, dynamic> stockTransactionHistoryData = {};
-    Map<String, dynamic> balanceData = {};
-    Map<String, dynamic> portfolioData = {};
 
     try {
       //await FirebaseFirestore.instance.collection('Users').add(datatoSave);
       DocumentReference userDocRef =
           await FirebaseFirestore.instance.collection('Users').add(datatoSave);
-
-      CollectionReference stockTransactionHistoryRef =
-          userDocRef.collection('stock_transaction_history');
-      await stockTransactionHistoryRef.add(stockTransactionHistoryData);
-
-      CollectionReference balanceHistoryRef =
-          userDocRef.collection('balance_history');
-      await balanceHistoryRef.add(balanceData);
-
-      CollectionReference portfolioRef = userDocRef.collection('portfolio');
-      await portfolioRef.add(portfolioData);
 
       // You can add additional logic or error handling here if needed
     } catch (e) {
@@ -52,9 +32,48 @@ class FirestoreService {
     }
   }
 
+  Future<num> getUserEmail() async {
+    User? user = FirebaseAuth.instance.currentUser;
+    if (user != null) {
+      final querySnapshot = await _firestore
+          .collection('Users')
+          .where('UID', isEqualTo: user?.uid)
+          .get();
+      final userDoc = querySnapshot.docs.first;
+      final userData = userDoc.data();
+      final balance = userData['balance'] as num;
+      return balance;
+    } else {
+      throw Exception("User is null, cannot retrieve balance.");
+    }
+  }
+
+  Future<num> getUserBalance() async {
+    User? user = FirebaseAuth.instance.currentUser;
+    if (user != null) {
+      final querySnapshot = await _firestore
+          .collection('Users')
+          .where('UID', isEqualTo: user?.uid)
+          .get();
+      final userDoc = querySnapshot.docs.first;
+      final userData = userDoc.data();
+      final balance = userData['balance'] as num;
+      return balance;
+    } else {
+      throw Exception("User is null, cannot retrieve balance.");
+    }
+  }
+
   //Es gibt keinen Befehl alle Subcollections auf einmal zu löschen, deswegen muss man es kompliziert machen
-  Future<void> deleteUser(String? documentID) async {
+  Future<bool> deleteUser(String? documentID, String password) async {
+    User? user = FirebaseAuth.instance.currentUser;
+    bool result;
     String? documentID = await getDocumentId();
+    num balance = await getUserBalance();
+
+    AuthCredential credentials =
+        EmailAuthProvider.credential(email: user!.email!, password: password);
+    await user.reauthenticateWithCredential(credentials);
 
     List<String> subcollections = [
       'stock_transaction_history',
@@ -62,18 +81,24 @@ class FirestoreService {
       'balance_history'
     ];
 
-    for (String subcollection in subcollections) {
-      var collectionRef = _firestore
-          .collection('Users')
-          .doc(documentID)
-          .collection(subcollection);
-      var snapshots = await collectionRef.get();
-      for (var doc in snapshots.docs) {
-        await doc.reference.delete();
+    if (balance == 0) {
+      for (String subcollection in subcollections) {
+        var collectionRef = _firestore
+            .collection('Users')
+            .doc(documentID)
+            .collection(subcollection);
+        var snapshots = await collectionRef.get();
+        for (var doc in snapshots.docs) {
+          await doc.reference.delete();
 
-        await _firestore.collection('Users').doc(documentID).delete();
+          await _firestore.collection('Users').doc(documentID).delete();
+        }
       }
+      result = true;
+    } else {
+      result = false;
     }
+    return result;
   }
 
   //Document ID /= User ID,deswegen brauchen wir ein Snapshot, um die DocumentID zu erhalten
